Remove ghg emission reducer. Use provider instead
